elyce_app/requirements.txt
Flask==2.0.1
pymongo==3.11.4
flask-pymongo==2.3.0
python-dotenv==0.19.0
gunicorn==20.0.4
Werkzeug==2.0.1

pytest
pytest-flask
Flask-Testing